Refactor Strength property to use double type and update bindings in LoraViewModel and LoraSelectionPage, Set proper mode for Z-Image Turbo LoRAs

+++ MobileDiffusion/ViewModels/LoraViewModel.cs
@@ -12,5 +12,5 @@ public partial class LoraViewModel : BaseViewModel, ILoraViewModel
     public partial string Name { get; set; }
 
     [ObservableProperty]
-    public partial float Strength { get; set; } = 1;
+    public partial double Strength { get; set; } = 1d;
 }
